style: customer 디렉터리 생성 및 관련 클래스 이동
