+++ src/main/java/woowacourse/customer/dto/ConfirmPasswordRequest.java
@@ -1,0 +1,17 @@
+package woowacourse.customer.dto;
+
+public class ConfirmPasswordRequest {
+
+    private String password;
+
+    public ConfirmPasswordRequest() {
+    }
+
+    public ConfirmPasswordRequest(final String password) {
+        this.password = password;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
