+++ src/main/java/woowacourse/customer/support/passwordencoder/PasswordEncoder.java
@@ -1,0 +1,6 @@
+package woowacourse.customer.support.passwordencoder;
+
+public interface PasswordEncoder {
+    String encode(String raw);
+    boolean matches(String raw, String encoded);
+}
