--- src/main/java/woowacourse/shoppingcart/dto/ConfirmPasswordRequest.java
@@ -1,17 +1,0 @@
-package woowacourse.shoppingcart.dto;
-
-public class ConfirmPasswordRequest {
-
-    private String password;
-
-    public ConfirmPasswordRequest() {
-    }
-
-    public ConfirmPasswordRequest(final String password) {
-        this.password = password;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
