--- src/main/java/woowacourse/shoppingcart/support/passwordencoder/PasswordEncoder.java
@@ -1,6 +1,0 @@
-package woowacourse.shoppingcart.support.passwordencoder;
-
-public interface PasswordEncoder {
-    String encode(String raw);
-    boolean matches(String raw, String encoded);
-}
